Adding more tests

+++ isshortenough.go
@@ -1,7 +1,7 @@
 package abbr
 
 // IsShortEnough checks if the short form is half or less in length compared to the long form.
-func IsShortEnough(long string, short string) bool {
+func IsShortEnough(short string, long string) bool {
 	if len(short)*2 > len(long) {
 		return false
 	}
